Remove the engines, just do DB clients
