+++ app/src/adapters/db/__init__.py
@@ -9,7 +9,7 @@ To use this module with Flask, use the flask_db module.
 Usage:
     import src.adapters.db as db
 
-    db_client = db.init(PostgresEngine())
+    db_client = db.init_postgres_client()
 
     # non-ORM style usage
     with db_client.get_connection() as conn:
@@ -23,9 +23,10 @@ Usage:
 """
 
 # Re-export for convenience
-from src.adapters.db.client import Connection, DBClient, Session, init
+from src.adapters.db.client import Connection, DBClient, Session
+from src.adapters.db.postgres_client import init_postgres_client
 
 # Do not import flask_db here, because this module is not dependent on any specific framework.
 # Code can choose to use this module on its own or with the flask_db module depending on needs.
 
-__all__ = ["Connection", "DBClient", "Session", "init"]
+__all__ = ["Connection", "DBClient", "Session", "init_postgres_client"]
