+++ app/tests/lib/db_testing.py
@@ -5,7 +5,6 @@ import uuid
 
 import src.adapters.db as db
 from src.adapters.db.config import get_db_config
-from src.adapters.db.engine.postgres_engine import PostgresEngine
 
 logger = logging.getLogger(__name__)
 
@@ -27,7 +26,7 @@ def create_isolated_db(monkeypatch) -> db.DBClient:
 
     # To improve test performance, don't check the database connection
     # when initializing the DB client.
-    db_client = db.init(PostgresEngine())
+    db_client = db.init_postgres_client()
     with db_client.get_connection() as conn:
         _create_schema(conn, schema_name)
         try:
